move agg nodes to separate package
