+++ modules/calcite/src/main/java/org/apache/ignite/internal/processors/query/calcite/rel/aggregate/IgniteMapSortAggregate.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.internal.processors.query.calcite.rel;
+package org.apache.ignite.internal.processors.query.calcite.rel.aggregate;
 
 import java.util.List;
 import java.util.Objects;
@@ -33,40 +33,49 @@ import org.apache.calcite.rel.RelWriter;
 import org.apache.calcite.rel.core.Aggregate;
 import org.apache.calcite.rel.core.AggregateCall;
 import org.apache.calcite.rel.metadata.RelMetadataQuery;
+import org.apache.calcite.rel.type.RelDataType;
+import org.apache.calcite.rel.type.RelDataTypeFactory;
+import org.apache.calcite.rel.type.RelDataTypeField;
 import org.apache.calcite.util.ImmutableBitSet;
 import org.apache.calcite.util.ImmutableIntList;
 import org.apache.calcite.util.Pair;
+import org.apache.ignite.internal.processors.query.calcite.exec.exp.agg.Accumulator;
+import org.apache.ignite.internal.processors.query.calcite.rel.IgniteConvention;
+import org.apache.ignite.internal.processors.query.calcite.rel.IgniteRel;
+import org.apache.ignite.internal.processors.query.calcite.rel.IgniteRelVisitor;
 import org.apache.ignite.internal.processors.query.calcite.trait.TraitUtils;
+import org.apache.ignite.internal.processors.query.calcite.util.Commons;
 
 import static org.apache.ignite.internal.processors.query.calcite.util.Commons.maxPrefix;
 
 /**
  *
  */
-public class IgniteSingleSortAggregate extends IgniteSingleAggregateBase {
+public class IgniteMapSortAggregate extends IgniteMapAggregateBase {
     /** Collation. */
     private final RelCollation collation;
 
-    /** {@inheritDoc} */
-    public IgniteSingleSortAggregate(
+    /** */
+    public IgniteMapSortAggregate(
         RelOptCluster cluster,
         RelTraitSet traitSet,
         RelNode input,
         ImmutableBitSet groupSet,
         List<ImmutableBitSet> groupSets,
-        List<AggregateCall> aggCalls
+        List<AggregateCall> aggCalls,
+        RelCollation collation
     ) {
         super(cluster, traitSet, input, groupSet, groupSets, aggCalls);
 
-        assert !TraitUtils.collation(traitSet).isDefault();
-        assert !groupSet.isEmpty() && groupSets.size() == 1;
+        assert Objects.nonNull(collation);
+        assert !collation.isDefault();
 
-        collation = TraitUtils.collation(traitSet);
+        this.collation = collation;
     }
 
-    /** {@inheritDoc} */
-    public IgniteSingleSortAggregate(RelInput input) {
-        super(input);
+    /** */
+    public IgniteMapSortAggregate(RelInput input) {
+        super(TraitUtils.changeTraits(input, IgniteConvention.INSTANCE));
 
         collation = input.getCollation();
 
@@ -75,14 +84,27 @@ public class IgniteSingleSortAggregate extends IgniteSingleAggregateBase {
     }
 
     /** {@inheritDoc} */
-    @Override public Aggregate copy(RelTraitSet traitSet, RelNode input, ImmutableBitSet groupSet, List<ImmutableBitSet> groupSets, List<AggregateCall> aggCalls) {
-        return new IgniteSingleSortAggregate(getCluster(), traitSet, input, groupSet, groupSets, aggCalls);
+    @Override public Aggregate copy(
+        RelTraitSet traitSet,
+        RelNode input,
+        ImmutableBitSet groupSet,
+        List<ImmutableBitSet> groupSets,
+        List<AggregateCall> aggCalls) {
+        return new IgniteMapSortAggregate(
+            getCluster(), traitSet, input, groupSet, groupSets, aggCalls, TraitUtils.collation(traitSet));
     }
 
     /** {@inheritDoc} */
     @Override public IgniteRel clone(RelOptCluster cluster, List<IgniteRel> inputs) {
-        return new IgniteSingleSortAggregate(cluster, getTraitSet(), sole(inputs),
-            getGroupSet(), getGroupSets(), getAggCallList());
+        return new IgniteMapSortAggregate(
+            cluster,
+            getTraitSet(),
+            sole(inputs),
+            getGroupSet(),
+            getGroupSets(),
+            getAggCallList(),
+            collation
+        );
     }
 
     /** {@inheritDoc} */
@@ -93,6 +115,33 @@ public class IgniteSingleSortAggregate extends IgniteSingleAggregateBase {
     /** {@inheritDoc} */
     @Override public RelWriter explainTerms(RelWriter pw) {
         return super.explainTerms(pw).item("collation", collation);
+    }
+
+    /** {@inheritDoc} */
+    @Override protected RelDataType deriveRowType() {
+        RelDataTypeFactory typeFactory = Commons.typeFactory(getCluster());
+
+        RelDataTypeFactory.Builder builder = new RelDataTypeFactory.Builder(typeFactory);
+
+        groupSet.forEach(fieldIdx -> {
+            RelDataTypeField fld = input.getRowType().getFieldList().get(fieldIdx);
+
+            builder.add(fld);
+        });
+
+        builder.add("AGG_DATA", typeFactory.createArrayType(typeFactory.createJavaType(Accumulator.class), -1));
+
+        return builder.build();
+    }
+
+    /** {@inheritDoc} */
+    @Override public RelOptCost computeSelfCost(RelOptPlanner planner, RelMetadataQuery mq) {
+        return computeSelfCostSort(planner, mq);
+    }
+
+    /** {@inheritDoc} */
+    @Override public RelCollation collation() {
+        return collation;
     }
 
     /** {@inheritDoc} */
@@ -120,17 +169,5 @@ public class IgniteSingleSortAggregate extends IgniteSingleAggregateBase {
             nodeTraits.replace(RelCollations.of(ImmutableIntList.copyOf(newCollation))),
             inputTraits
         ));
-    }
-
-    /** {@inheritDoc} */
-    @Override public RelCollation collation() {
-        assert collation.equals(super.collation());
-
-        return collation;
-    }
-
-    /** {@inheritDoc} */
-    @Override public RelOptCost computeSelfCost(RelOptPlanner planner, RelMetadataQuery mq) {
-        return computeSelfCostSort(planner, mq);
     }
 }
