+++ modules/calcite/src/main/java/org/apache/ignite/internal/processors/query/calcite/rel/aggregate/IgniteSingleSortAggregate.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.internal.processors.query.calcite.rel;
+package org.apache.ignite.internal.processors.query.calcite.rel.aggregate;
 
 import java.util.List;
 import java.util.Objects;
@@ -33,47 +33,42 @@ import org.apache.calcite.rel.RelWriter;
 import org.apache.calcite.rel.core.Aggregate;
 import org.apache.calcite.rel.core.AggregateCall;
 import org.apache.calcite.rel.metadata.RelMetadataQuery;
-import org.apache.calcite.rel.type.RelDataType;
-import org.apache.calcite.rel.type.RelDataTypeFactory;
-import org.apache.calcite.rel.type.RelDataTypeField;
 import org.apache.calcite.util.ImmutableBitSet;
 import org.apache.calcite.util.ImmutableIntList;
 import org.apache.calcite.util.Pair;
-import org.apache.ignite.internal.processors.query.calcite.exec.exp.agg.Accumulator;
+import org.apache.ignite.internal.processors.query.calcite.rel.IgniteRel;
+import org.apache.ignite.internal.processors.query.calcite.rel.IgniteRelVisitor;
 import org.apache.ignite.internal.processors.query.calcite.trait.TraitUtils;
-import org.apache.ignite.internal.processors.query.calcite.util.Commons;
 
-import static org.apache.ignite.internal.processors.query.calcite.trait.TraitUtils.changeTraits;
 import static org.apache.ignite.internal.processors.query.calcite.util.Commons.maxPrefix;
 
 /**
  *
  */
-public class IgniteMapSortAggregate extends IgniteMapAggregateBase {
+public class IgniteSingleSortAggregate extends IgniteSingleAggregateBase {
     /** Collation. */
     private final RelCollation collation;
 
-    /** */
-    public IgniteMapSortAggregate(
+    /** {@inheritDoc} */
+    public IgniteSingleSortAggregate(
         RelOptCluster cluster,
         RelTraitSet traitSet,
         RelNode input,
         ImmutableBitSet groupSet,
         List<ImmutableBitSet> groupSets,
-        List<AggregateCall> aggCalls,
-        RelCollation collation
+        List<AggregateCall> aggCalls
     ) {
         super(cluster, traitSet, input, groupSet, groupSets, aggCalls);
 
-        assert Objects.nonNull(collation);
-        assert !collation.isDefault();
+        assert !TraitUtils.collation(traitSet).isDefault();
+        assert !groupSet.isEmpty() && groupSets.size() == 1;
 
-        this.collation = collation;
+        collation = TraitUtils.collation(traitSet);
     }
 
-    /** */
-    public IgniteMapSortAggregate(RelInput input) {
-        super(changeTraits(input, IgniteConvention.INSTANCE));
+    /** {@inheritDoc} */
+    public IgniteSingleSortAggregate(RelInput input) {
+        super(input);
 
         collation = input.getCollation();
 
@@ -82,27 +77,14 @@ public class IgniteMapSortAggregate extends IgniteMapAggregateBase {
     }
 
     /** {@inheritDoc} */
-    @Override public Aggregate copy(
-        RelTraitSet traitSet,
-        RelNode input,
-        ImmutableBitSet groupSet,
-        List<ImmutableBitSet> groupSets,
-        List<AggregateCall> aggCalls) {
-        return new IgniteMapSortAggregate(
-            getCluster(), traitSet, input, groupSet, groupSets, aggCalls, TraitUtils.collation(traitSet));
+    @Override public Aggregate copy(RelTraitSet traitSet, RelNode input, ImmutableBitSet groupSet, List<ImmutableBitSet> groupSets, List<AggregateCall> aggCalls) {
+        return new IgniteSingleSortAggregate(getCluster(), traitSet, input, groupSet, groupSets, aggCalls);
     }
 
     /** {@inheritDoc} */
     @Override public IgniteRel clone(RelOptCluster cluster, List<IgniteRel> inputs) {
-        return new IgniteMapSortAggregate(
-            cluster,
-            getTraitSet(),
-            sole(inputs),
-            getGroupSet(),
-            getGroupSets(),
-            getAggCallList(),
-            collation
-        );
+        return new IgniteSingleSortAggregate(cluster, getTraitSet(), sole(inputs),
+            getGroupSet(), getGroupSets(), getAggCallList());
     }
 
     /** {@inheritDoc} */
@@ -113,33 +95,6 @@ public class IgniteMapSortAggregate extends IgniteMapAggregateBase {
     /** {@inheritDoc} */
     @Override public RelWriter explainTerms(RelWriter pw) {
         return super.explainTerms(pw).item("collation", collation);
-    }
-
-    /** {@inheritDoc} */
-    @Override protected RelDataType deriveRowType() {
-        RelDataTypeFactory typeFactory = Commons.typeFactory(getCluster());
-
-        RelDataTypeFactory.Builder builder = new RelDataTypeFactory.Builder(typeFactory);
-
-        groupSet.forEach(fieldIdx -> {
-            RelDataTypeField fld = input.getRowType().getFieldList().get(fieldIdx);
-
-            builder.add(fld);
-        });
-
-        builder.add("AGG_DATA", typeFactory.createArrayType(typeFactory.createJavaType(Accumulator.class), -1));
-
-        return builder.build();
-    }
-
-    /** {@inheritDoc} */
-    @Override public RelOptCost computeSelfCost(RelOptPlanner planner, RelMetadataQuery mq) {
-        return computeSelfCostSort(planner, mq);
-    }
-
-    /** {@inheritDoc} */
-    @Override public RelCollation collation() {
-        return collation;
     }
 
     /** {@inheritDoc} */
@@ -167,5 +122,17 @@ public class IgniteMapSortAggregate extends IgniteMapAggregateBase {
             nodeTraits.replace(RelCollations.of(ImmutableIntList.copyOf(newCollation))),
             inputTraits
         ));
+    }
+
+    /** {@inheritDoc} */
+    @Override public RelCollation collation() {
+        assert collation.equals(super.collation());
+
+        return collation;
+    }
+
+    /** {@inheritDoc} */
+    @Override public RelOptCost computeSelfCost(RelOptPlanner planner, RelMetadataQuery mq) {
+        return computeSelfCostSort(planner, mq);
     }
 }
